requirements.txt
allure-pytest==2.13.5
allure_python_commons==2.13.5
Flask==2.2.5
loguru==0.7.2
pluggy==1.2.0
pycryptodome==3.20.0
pycryptodomex==3.7.2
pytest==7.4.4
python-jenkins==1.8.2
PyYAML==6.0.1
requests==2.21.0
urllib3==1.24.3
urllib3-secure-extra==0.1.0
xlrd==2.0.1
zipp==3.15.0